Services/api_temperature/insertData/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
mysql-connector-python==8.0.33
protobuf==3.20.3
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.3
